refactor(factory): migrate SlackIntegration to Messaging capability and remove deprecated channels() slot
- Extract account-linking gate helpers to integrations/slack/gates.ts:
  createGithubSourceControl, rawTeamId, SlackSourceControl,
  resolveLinkedSender, resolveFactoryForLink, gateDispatch, and new
  resolveWorkspaceContext(accountLinks, ref) shared by both Slack
  integrations. Behavior-preserving move.

- SlackIntegration now owns a Messaging capability built in the constructor.
  Private #buildChannels(ctx) returns the config; #resolveWorkspaceContext
  delegates to gates.resolveWorkspaceContext. Top-level channels(ctx) removed.

- Remove deprecated FactoryIntegration.channels?() slot. Factory readiness
  gate and channel-contribution collection now consult only messaging.

- Update Slack tests to drive integration.messaging.channels/
  resolveWorkspaceContext; move gate-helper imports to ./gates.js. Remove
  the Phase 2 migration-shim test that combined messaging with the
  deprecated channels slot.

Verification: factory tests (164 pass), factory typecheck, factory-ui
typecheck, factory lint, core channels/agent-controller tests (674 pass),
oxfmt --check.
