+++ mastracode/factory/src/integrations/slack/slack.ts
@@ -1,24 +1,14 @@
-import { randomUUID } from 'node:crypto';
-
-import type {
-  ChannelHandler,
-  ChannelHandlerContext,
-  ChannelHandlers,
-  ResolveResourceId,
-  ResolveThreadId,
-} from '@mastra/core/channels';
+import type { ChannelHandler, ChannelHandlers, ResolveResourceId, ResolveThreadId } from '@mastra/core/channels';
 import type { Mastra } from '@mastra/core/mastra';
 import { createSlackAdapter } from '@mastra/slack';
-import { Card, CardText, Actions, LinkButton } from 'chat';
+import { Card, Actions, LinkButton } from 'chat';
 
-import type {
-  ChannelAccountLink,
-  ChannelAccountLinkKey,
-  ChannelIdentityStorage,
-} from '../../storage/domains/channel-identity/base.js';
+import type { ChannelAccountLink, ChannelIdentityStorage } from '../../storage/domains/channel-identity/base.js';
 import type { FactoryProjectsStorage } from '../../storage/domains/projects/base.js';
 import type { WorkItemsStorage } from '../../storage/domains/work-items/base.js';
 import type { FactoryChannelsConfig } from '../base.js';
+
+import { gateDispatch, rawTeamId, type SlackSourceControl } from './gates.js';
 
 // Derive the thread/message types from the core handler signature rather than
 // importing them from `chat` directly: mc-web can resolve a different `chat`
@@ -62,283 +52,6 @@ interface SlackChannelDeps {
    * session. Best-effort — a failure never blocks the run. Unset → no card.
    */
   workItems?: WorkItemsStorage;
-}
-
-/**
- * The slice of the GitHub integration's source-control storage the Slack
- * wiring needs. Structural (not the storage types themselves) so slack.ts
- * stays decoupled from the integration's module graph and tests can stub it.
- */
-export interface SlackSourceControl {
-  /**
-   * Resolve the factory's linked repository — first repo on the factory's
-   * GitHub connection, the same single-repo assumption the web kickoff makes.
-   */
-  resolveProjectRepository(args: {
-    orgId: string;
-    factoryProjectId: string;
-  }): Promise<{ projectRepositoryId: string; baseBranch: string } | null>;
-  /** Look up an existing user session for a repo branch (idempotent reuse). */
-  getSessionForBranch(args: {
-    projectRepositoryId: string;
-    userId: string;
-    branch: string;
-  }): Promise<{ sessionId: string } | null>;
-  /** Create the durable user-session row the workspace factory materializes. */
-  createSession(args: {
-    projectRepositoryId: string;
-    orgId: string;
-    userId: string;
-    branch: string;
-    baseBranch: string;
-  }): Promise<{ sessionId: string }>;
-}
-
-/**
- * Structural slice of the source-control owner's storage handle
- * (`IntegrationContext.storage.sourceControlOwner`, a `SourceControlStorageHandle`)
- * the adapter reads. Structural so slack.ts stays decoupled from the storage
- * module graph and tests can stub it. The handle is bound during
- * `factory.prepare()`; all access here is lazy (request time).
- */
-interface SourceControlOwnerSlice {
-  integrationId: string;
-  connections: {
-    list(args: { orgId: string; factoryProjectId: string }): Promise<Array<{ id: string; integrationId: string }>>;
-  };
-  projectRepositories: {
-    list(args: {
-      orgId: string;
-      connectionId: string;
-    }): Promise<Array<{ id: string; repositoryId: string; branch?: string | null }>>;
-  };
-  repositories: {
-    get(args: { orgId: string; id: string }): Promise<{ defaultBranch: string } | null>;
-  };
-  sessions: {
-    getForBranch(args: {
-      projectRepositoryId: string;
-      userId: string;
-      branch: string;
-    }): Promise<{ sessionId: string } | null>;
-    create(input: {
-      sessionId: string;
-      projectRepositoryId: string;
-      orgId: string;
-      userId: string;
-      branch: string;
-      baseBranch: string;
-    }): Promise<{ sessionId: string }>;
-  };
-}
-
-/**
- * Adapt the source-control owner's storage handle into the
- * {@link SlackSourceControl} surface. Repo resolution mirrors the factory's
- * own `ensureFactoryRuleSession`: the factory's GitHub connection → its first
- * linked repository → pinned branch or repo default as the base.
- */
-export function createGithubSourceControl(owner: SourceControlOwnerSlice): SlackSourceControl {
-  return {
-    async resolveProjectRepository({ orgId, factoryProjectId }) {
-      const connections = await owner.connections.list({ orgId, factoryProjectId });
-      const connection = connections.find(candidate => candidate.integrationId === owner.integrationId);
-      if (!connection) return null;
-      const projectRepositories = await owner.projectRepositories.list({ orgId, connectionId: connection.id });
-      const first = projectRepositories[0];
-      if (!first) return null;
-      const repository = await owner.repositories.get({ orgId, id: first.repositoryId });
-      if (!repository) return null;
-      return { projectRepositoryId: first.id, baseBranch: first.branch ?? repository.defaultBranch };
-    },
-    getSessionForBranch: args => owner.sessions.getForBranch(args),
-    createSession: args => owner.sessions.create({ sessionId: randomUUID(), ...args }),
-  };
-}
-
-/**
- * Read the Slack team id off a raw platform payload (Events API envelope or
- * slash-command body — both carry `team_id`), duck-typed to build the
- * workspace-scoped account-link key.
- */
-function rawTeamId(rawPayload: unknown): string | undefined {
-  if (!rawPayload || typeof rawPayload !== 'object') return undefined;
-  const raw = rawPayload as { team_id?: unknown; team?: unknown };
-  if (typeof raw.team_id === 'string' && raw.team_id) return raw.team_id;
-  if (typeof raw.team === 'string' && raw.team) return raw.team;
-  if (raw.team && typeof raw.team === 'object') {
-    const id = (raw.team as { id?: unknown }).id;
-    if (typeof id === 'string' && id) return id;
-  }
-  return undefined;
-}
-
-/**
- * The Slack team id survives onto a normalized chat Message only on
- * `message.raw` (the Slack Events API envelope).
- */
-function slackTeamId(message: HandlerMessage): string | undefined {
-  return rawTeamId(message.raw);
-}
-
-/**
- * Resolve the web-UI origin for links humans open in a browser (Connect card,
- * session deep links). Prefers `MASTRACODE_PUBLIC_URL` — the origin auth
- * cookies and OAuth redirect allow-lists are registered against — over the
- * channels tunnel, which only Slack's servers need to reach.
- */
-function webPublicUrl(): string | undefined {
-  return process.env.MASTRACODE_PUBLIC_URL ?? process.env.MASTRACODE_CHANNELS_PUBLIC_URL;
-}
-
-/** Outcome of the sender-link gate for one inbound message. */
-type LinkedSenderResult =
-  /** Gating not configured — dispatch as before account linking existed. */
-  | { status: 'ungated' }
-  /** Sender unlinked — Connect card posted (when possible), do not dispatch. */
-  | { status: 'blocked' }
-  /** Sender linked — their tenant plus the sender key the link lives under. */
-  | { status: 'linked'; link: ChannelAccountLink; key: ChannelAccountLinkKey };
-
-/**
- * Resolve the sender's account link, posting an ephemeral "connect your
- * account" card (visible only to the sender) linking into the web UI's
- * Slack-connect flow when they're unlinked.
- */
-export async function resolveLinkedSender({
-  thread,
-  message,
-  accountLinks,
-}: {
-  thread: HandlerThread;
-  message: HandlerMessage;
-  accountLinks?: ChannelIdentityStorage;
-}): Promise<LinkedSenderResult> {
-  if (!accountLinks) return { status: 'ungated' };
-  const platform = thread.adapter.name;
-  const externalUserId = message.author.userId;
-  const externalTeamId = slackTeamId(message);
-  // Without a team id we can't identify the workspace-scoped link; treat as
-  // unlinked so a run never proceeds tenant-less.
-  const key = externalTeamId ? { platform, externalTeamId, externalUserId } : undefined;
-  const link = key ? await accountLinks.getAccountLink(key) : null;
-  if (link && key) return { status: 'linked', link, key };
-
-  const publicUrl = webPublicUrl();
-  // A public origin is all the card needs. The link carries no identity: the
-  // web app authenticates the visitor, then Slack's OIDC flow proves which
-  // Slack account they control. Without an origin, still block, just no card.
-  if (publicUrl) {
-    await thread.postEphemeral(message.author, buildConnectCard(publicUrl), { fallbackToDM: true });
-  }
-  return { status: 'blocked' };
-}
-
-/**
- * The "connect your account" card. The link is deliberately identity-free —
- * `/connect/slack` sends the visitor to Connections, where "Connect Slack"
- * runs the OIDC flow and Slack itself asserts the (team, user) pair.
- */
-function buildConnectCard(publicUrl: string) {
-  return Card({
-    title: 'Connect your account',
-    children: [
-      CardText('Connect your account to use this agent.'),
-      Actions([
-        LinkButton({
-          url: `${publicUrl}/connect/slack`,
-          label: 'Connect account',
-        }),
-      ]),
-    ],
-  });
-}
-
-/** Outcome of factory routing for one linked sender's inbound message. */
-type FactoryRouteResult =
-  /** Factory routing not configured — dispatch without a factory. */
-  | { status: 'ungated' }
-  /** No factory resolved — prompt card posted (when possible), do not dispatch. */
-  | { status: 'blocked' }
-  /** The Factory project this sender's runs route to. */
-  | { status: 'resolved'; factoryProjectId: string; slackWorkItemsEnabled: boolean };
-
-/**
- * Decide which Factory project a linked sender's run belongs to:
- *
- * 1. The link's `defaultFactoryProjectId`, when it still exists (a stale id —
- *    deleted factory — falls through as if unset).
- * 2. Else, the tenant's only factory, stamped back onto the link so it shows
- *    up (and stays editable) in Connected Accounts settings.
- * 3. Else — zero or several factories — an ephemeral "pick a default factory"
- *    card deep-linking to settings, and the run is blocked.
- */
-export async function resolveFactoryForLink({
-  thread,
-  message,
-  link,
-  key,
-  accountLinks,
-  projects,
-}: {
-  thread: HandlerThread;
-  message: HandlerMessage;
-  link: ChannelAccountLink;
-  key: ChannelAccountLinkKey;
-  accountLinks: ChannelIdentityStorage;
-  projects?: FactoryProjectsStorage;
-}): Promise<FactoryRouteResult> {
-  if (!projects) return { status: 'ungated' };
-  // Factories are org-scoped; a personal account (no org) has none and lands
-  // on the prompt below.
-  const orgId = link.orgId ?? '';
-
-  if (link.defaultFactoryProjectId) {
-    const existing = await projects.get({ orgId, id: link.defaultFactoryProjectId });
-    if (existing) {
-      return {
-        status: 'resolved',
-        factoryProjectId: existing.id,
-        slackWorkItemsEnabled: existing.slackWorkItemsEnabled,
-      };
-    }
-  }
-
-  const factories = orgId ? await projects.list({ orgId }) : [];
-  if (factories.length === 1) {
-    const only = factories[0]!;
-    await accountLinks.setDefaultFactory({ ...key, userId: link.userId, factoryProjectId: only.id });
-    return {
-      status: 'resolved',
-      factoryProjectId: only.id,
-      slackWorkItemsEnabled: only.slackWorkItemsEnabled,
-    };
-  }
-
-  const publicUrl = webPublicUrl();
-  if (publicUrl) {
-    await thread.postEphemeral(
-      message.author,
-      Card({
-        title: 'Pick a default factory',
-        children: [
-          CardText(
-            factories.length === 0
-              ? 'Your account has no factory yet. Create one in the web app, then message me again.'
-              : 'Your account has several factories. Pick which one Slack sessions should go to, then message me again.',
-          ),
-          Actions([
-            LinkButton({
-              url: `${publicUrl}/settings/connections`,
-              label: 'Open settings',
-            }),
-          ]),
-        ],
-      }),
-      { fallbackToDM: true },
-    );
-  }
-  return { status: 'blocked' };
 }
 
 /**
@@ -461,47 +174,6 @@ async function findInternalThread(mastra: Mastra | undefined, thread: HandlerThr
  * DM on a not-yet-subscribed thread starts a NEW session; once subscribed, later
  * events are follow-ups and don't re-announce.
  */
-/**
- * Run the account-link + factory-routing gates for one inbound message.
- * Returns `null` when the run must not dispatch (a prompt card was posted
- * where possible); otherwise the dispatch context — with `routed` present
- * only when a linked sender resolved to a factory.
- */
-async function gateDispatch(
-  thread: HandlerThread,
-  message: HandlerMessage,
-  { accountLinks, projects }: SlackChannelDeps,
-  ctx: ChannelHandlerContext,
-): Promise<{
-  routed?: { link: ChannelAccountLink; factoryProjectId: string; slackWorkItemsEnabled: boolean };
-} | null> {
-  const sender = await resolveLinkedSender({ thread, message, accountLinks });
-  if (sender.status === 'blocked') return null;
-  // Linked senders must also route to a Factory project before a run starts.
-  if (sender.status === 'linked' && accountLinks) {
-    // Stamp the tenant on the run's request context — the single seam
-    // `resolveCredentialStore` reads to load this sender's model credentials.
-    // This belongs to the link, not to the routing: a linked sender whose
-    // factory routing comes back `ungated` still exits below and dispatches, so
-    // stamping only in the routed branch would silently run them on default
-    // credentials.
-    ctx.requestContext.set('user', { id: sender.link.userId, organizationId: sender.link.orgId });
-
-    const route = await resolveFactoryForLink({ thread, message, ...sender, accountLinks, projects });
-    if (route.status === 'blocked') return null;
-    if (route.status === 'resolved') {
-      return {
-        routed: {
-          link: sender.link,
-          factoryProjectId: route.factoryProjectId,
-          slackWorkItemsEnabled: route.slackWorkItemsEnabled,
-        },
-      };
-    }
-  }
-  return {};
-}
-
 /**
  * Upsert the Work-board card for a dispatched Slack-thread run. Keyed on the
  * thread via `externalSource` — the work-items domain's unique
